services/ml-service/requirements.txt
fastapi>=0.85.1
matplotlib>=3.6.1
# numpy==1.23.3
numpy<2.0.0
opencv_python>=4.6.0.66
Pillow>=9.2.0
torch>=1.12.1
uvicorn>=0.19.0

# YOLOv5 ----------------------------------------
matplotlib>=3.2.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torchvision>=0.8.1
tqdm>=4.64.0
python-multipart==0.0.5

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb
# clearml

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=6.0  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn<=1.1.2  # CoreML quantization
# tensorflow>=2.4.1  # TF exports (-cpu, -aarch64, -macos)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Deploy --------------------------------------
# tritonclient[all]~=2.24.0

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
# mss  # screenshots
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
